better version range resolver
gradle just picks the latest and newest version when there are conflicting
versions ranges regardless whether it obeys the involved ranges or not.

this patch tries obey ranges. if there is no overlap between two ranges for the
same gem then it fails. the second range the same gem will be narrowed down to
the intersection of the first and second range.

it still can fail but now you can fix things be reordering you dependencies a bit
for most cases.