run version resolution twice to fix ALL version ranges
only when the gem get resolved first and the version ranges get collected those
info can be applied to ALL versions ranges when the same set of gem dependencies
get reolved a second time.

it comes with test BUT this test would need a live repo with gem artifacts since the
flatRepo does not provide info about transitive dependencies which are the actual point
here. the gems are choosen to produce a false result prior to this patch.