require jar dependencies instead of attaching them to the classpath
basically it sets up the jars in the same manner as it would pack it inside a jar
and uses the usual loading semantic for loading jars with jruby. and with this
the jrubyExec task does not need interfer with those embedded jars coming with
jruby itself like bouncycastle jars, jline jar or snakeyaml jar.

fixes #123 and #113